metrics2 -> metrics (#270)
