--- controller/metrics2/config.go
@@ -1,17 +1,0 @@
-package metrics2
-
-type Config struct {
-	Influx *InfluxConfig
-	Agent  *AgentConfig
-}
-
-type AgentConfig struct {
-	Source interface{}
-}
-
-type InfluxConfig struct {
-	Url    string
-	Bucket string
-	Org    string
-	Token  string `cf:"+secret"`
-}
